See #2287: The data access should not throw WebException with HTTP Status code 404
Remove EntityUtil.populateOwner

Remove ParseException from the code where it was not necessary.

Remove EntityTypeNotValidException. Rename EntityTypeNotValidCheckedExeception to EntityTypeNotFoundExeception.

Remove EntityTypeNoFoundException and EntityNotFoundCheckedException from all signatures because it was not needed. Renamed ValidationException to BadRequestException.

Add a test of inconsistency of the database (a table doesn't have a database link). Fix a typo in CatalogExceptionMessage.

See #2331: Use a fluent API with the decorator

Better logs and correct semantic of getContainer.

Add danglingLink log msg.

Remove FIXME.

Better documentation for beforeDeletion.