Fix facet filter issue
Issue was the following: the user set a facet filter on `aggregationKey` and `value`. Then user changes the query builder so that in the new result there is no bucket in `aggregationKey` with `value`. The facet filter would still be used, but not visible to the user and not removable.