UI (#6189) | Refactor Explore and ExplorePage, add Logical Query Builder (#8140)
* Add logical query builder and Explore page refactor

* Add file headers

* Change search aggreagation keys

* Move query builder utils, remove comments

* Add comments

* Rename AdvancesSearch.constants.ts

* Add review changes

* Handle deleted elements correctly in Explore

The `deleted` field needs special handling because `showDeleted === false` should return only documents where `deleted` is false, but `showDeleted === true` should return documents regardless of their `deleted` field.

* Fix facet filter issue

Issue was the following: the user set a facet filter on `aggregationKey` and `value`. Then user changes the query builder so that in the new result there is no bucket in `aggregationKey` with `value`. The facet filter would still be used, but not visible to the user and not removable.

* Fix facetFilter tests

* Move isFilterObject type guard

* Use antd tabs for Explore component

* fix cypress failure

* miner fix in cypress

* fix typo

Co-authored-by: Chirag Madlani <12962843+chirag-madlani@users.noreply.github.com>
Co-authored-by: Shailesh Parmar <shailesh.parmar.webdev@gmail.com>