Prevent sending email field with an empty string value in the HTTP POST request for creating a team. (#12148)
Update behavior to disallow submitting an empty string value for the email field when creating a team. If the email input is left blank without entering any characters, no problem is encountered. However, if the email input is filled and then emptied, an exception with the message "[email must match "^\S+@\S+.\S+$", email size must be between 6 and 127]" was previously thrown because the email pattern did not accept empty strings. Rather than altering the email pattern to accept empty strings, the change ensures that empty string values are not submitted.

Co-authored-by: Chirag Madlani <12962843+chirag-madlani@users.noreply.github.com>