Refactor date formatting in KPIChart and DataQualityUtils
This commit refactors the date formatting in the KPIChart component and the DataQualityUtils module. In the KPIChart component, the tickFormatter function in the XAxis component is updated to use an arrow function instead of the formatDate function. In the DataQualityUtils module, the formatDate function is modified to support UTC formatting when necessary. Additionally, two functions, getStartOfDayInMillis and getEndOfDayInMillis, are added to the DateTimeUtils module to calculate the start and end of a day in milliseconds.