ISSUE-16094: fix s3 storage parquet structureFormat ingestion
This aims at fixing the s3 storage ingestion for parquet structured containers,
when the ingestion flow tries to extract columns from a sample file under the dataPath
it may happen that a non structureFormat file is picked, example a _SUCCESS file or
that the prefix is empty but the s3 lifecycle hasnt kicked in yet, this will result in
the column extraction understandably failing as the file cannot be loaded into a dataframe.

For that, do not fail when the column extraction fails but log a warning and continue with
the remainder of the paths in the manifest, this is already an improvement on the status quo,
but ideally when the column extraction fails, the ingestion flow should have a rety mechanism
where it will rety a couple other files under the prefix if available before giving up, and moving
on to another prefix.

Refrence issues: