Improvements: Automator requirements oss (#19128)
* Fix the ingestion success screen text overflow issue

* Fix the icon alignment issue in the in the entity header

* Fix the permissions API calls for the ingestion pipelines to use fqn instead of names

* Introduce the moreActionButtonProps in the PipelineActions component

* Add the "Database" field in the advanced search for the "Database Schema" asset.
Change the fields listing behavior in the advanced search modal, to show the common fields for the multiple assets instead of union of the fields.
Change the search index for the "Name" and "Display Name" field suggestions to use the selected data assets instead of the all "Data Assets" index.

* Remove the unnecessary ellipsis prop

* Scan and add the missing filters for all the asset to have hierarchical parent filters.

* Fix unit test

* Fix the advanced search filter field search