Enable sorting by `_score`
Sorting by `_score` would break the search, since it results in deserialization error.
The reason is, the server could return either a list of strings, or a list of
floats and strings, depends on the filter field.
When the filter field is _score, the ranking score is returned, causing pydantic
to yell.

I chose to go with a simple convert-all-items-to-string approach, because this
data comes from a trusted open api server and not from an end user. It's about
data representation more than actual validation.